+++ src/common/Pagination/Arrow/nextArrow.js
@@ -1,0 +1,5 @@
+import { RightArrow } from "./styled";
+
+export const NextArrow = ({ disabled }) => {
+  return <RightArrow disabled={disabled} />;
+};
