+++ src/common/Pagination/Arrow/previousArrow.js
@@ -1,0 +1,5 @@
+import { LeftArrow } from "./styled";
+
+export const PreviousArrow = ({ disabled }) => {
+  return <LeftArrow disabled={disabled} />;
+};
