--- src/common/Pagination/NextArrow/index.js
@@ -1,5 +1,0 @@
-import { Arrow } from "./styled";
-
-export const NextArrow = ({ disabled }) => {
-  return <Arrow disabled={disabled} />;
-};
