--- src/common/Pagination/PreviousArrow/index.js
@@ -1,5 +1,0 @@
-import { Arrow } from "./styled";
-
-export const PreviousArrow = ({ disabled }) => {
-  return <Arrow disabled={disabled} />;
-};
